convert UserOnCallAssignmentList to TS
